add focused window geometry history operations.
Implement a way of obtaining the history of all windows of the same app
in a visibility context and traversing the unique geometries in that
history in either direction for the focused window.

This is something I've been thinking about, and this is a first step to
experiment with.  In this case you can easily take one window backward
and forward through the history of geometries.  It may be the case that
I want to move a group through together, not sure.